+++ metadata.rb
@@ -4,4 +4,4 @@ maintainer_email 'edwardrsmi@gmail.com'
 license          'MIT'
 description      'Installs and configures openssh-client and openssh-server.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.1'
+version          '0.3.0'