+++ recipes/ssh_client.rb
@@ -1,4 +1,0 @@
-# Installs the openssh client
-package node[:ssh][:client][:package_name] do
-  action :install
-end